Notify when 'sudo' is not installed
Related: #385

The script is providing the wrong error feedback (saying it could not find
the configuration directory) when run on Docker. The actual issue is
that it cannot locate a suitable `sudo` command. While there is not
specific advice for Docker users, this will at least clear up some
confusion (potentially for other environments where `sudo` is
purposefully removed) with the `install_demo_configuration.sh` script.